Refactor: index.ts 기반으로 api import 방식 변경
